removing dead code. Moved querystresstestvm into testenv namespace.
