--- SqlQueryStressGUI/QueryStressTests/QueryExecutionResultsGridViewModel.cs
@@ -1,10 +1,0 @@
-﻿using SqlQueryStressEngine;
-using System.Collections.ObjectModel;
-
-namespace SqlQueryStressGUI.QueryStressTests
-{
-    public class QueryExecutionResultsGridViewModel : ViewModel
-    {
-        public ObservableCollection<QueryExecutionStatistics> Results { get; set; }
-    }
-}
