--- SqlQueryStressGUI/QueryStressTests/Views/QueryStressTestPage.xaml
@@ -1,124 +1,0 @@
-﻿<Page x:Class="SqlQueryStressGUI.QueryStressTests.Views.QueryStressTestPage"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-      xmlns:designContexts="clr-namespace:SqlQueryStressGUI.DesignTimeContexts"
-      xmlns:views="clr-namespace:SqlQueryStressGUI.QueryStressTests.Views"
-      xmlns:teviews="clr-namespace:SqlQueryStressGUI.TestEnvironment.Views"
-      xmlns:i="clr-namespace:System.Windows.Interactivity;assembly=System.Windows.Interactivity"
-      xmlns:ext="clr-namespace:System.Windows.Controls.Extensions"
-      xmlns:sqs="clr-namespace:SqlQueryStressGUI"
-      d:DataContext="{x:Static designContexts:DesignContexts.QueryStressTestContext}"
-      mc:Ignorable="d" 
-      d:DesignHeight="450" d:DesignWidth="1000"
-      x:Name="queryStressTestPage"
-      Title="QueryStressTestPage">
-    <Grid>
-        <Grid.Resources>
-            <sqs:BindingProxy x:Key="Proxy" Data="{Binding}" />
-        </Grid.Resources>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="35"/>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="35"/>
-        </Grid.RowDefinitions>
-
-        <ToolBarTray Grid.Row="0">
-            <ToolBar>
-                <StackPanel Margin="5,0,0,0" Orientation="Horizontal">
-                    <TextBlock Text="Connection" VerticalAlignment="Center" />
-                    <ComboBox
-                        Grid.Column="1"
-                        ItemsSource="{Binding Path=Connections}"
-                        SelectedItem="{Binding Path=SelectedConnection}"
-                        DisplayMemberPath="Name"
-                        Padding="5"
-                        Width="150"
-                        Margin="5,0,0,0">
-                        <i:Interaction.Triggers>
-                            <i:EventTrigger EventName="DropDownClosed">
-                                <i:InvokeCommandAction Command="{Binding ConnectionDropdownClosedCommand}"/>
-                            </i:EventTrigger>
-                            <i:EventTrigger EventName="SelectionChanged">
-                                <i:InvokeCommandAction Command="{Binding ConnectionChangedCommand}"/>
-                            </i:EventTrigger>
-                        </i:Interaction.Triggers>
-                    </ComboBox>
-                </StackPanel>
-
-                <StackPanel Orientation="Horizontal" Margin="10,0,0,0">
-                    <TextBlock Text="Threads" HorizontalAlignment="Center" VerticalAlignment="Center" Margin="0,0,10,0" />
-                    <TextBox Text="{Binding Path=ThreadCount}" Padding="5" PreviewTextInput="TextBox_PreviewTextInput" />
-                </StackPanel>
-
-                <StackPanel Orientation="Horizontal" Margin="10,0,0,0">
-                    <TextBlock Text="Iterations" HorizontalAlignment="Center" VerticalAlignment="Center" Margin="0,0,10,0" />
-                    <TextBox Text="{Binding Path=Iterations}" Padding="5" PreviewTextInput="TextBox_PreviewTextInput" />
-                </StackPanel>
-                
-                <Separator Margin="5,0,5,0" Style="{StaticResource {x:Static ToolBar.SeparatorStyleKey}}" />
-
-                <Button Margin="0,0,5,0" Command="{Binding GoCommandHandler}" CommandParameter="{Binding ElementName=queryEditor, Path=SelectedQuery}">
-                    <StackPanel Orientation="Horizontal">
-                        <Rectangle Height="18" Width="16" Fill="Green" >
-                            <Rectangle.OpacityMask>
-                                <VisualBrush Stretch="Fill" Visual="{StaticResource appbar_control_play}" />
-                            </Rectangle.OpacityMask>
-                        </Rectangle>
-                        <TextBlock Margin="5,0,0,0" Text="Execute" VerticalAlignment="Center" />
-                    </StackPanel>
-                </Button>
-
-                <Separator Margin="5,0,5,0" Style="{StaticResource {x:Static ToolBar.SeparatorStyleKey}}" />
-
-                <ext:DropdownButton>
-                    <ext:DropdownButton.Content>
-                        Database Commands
-                    </ext:DropdownButton.Content>
-                    <ext:DropdownButton.DropdownMenu>
-                        <ContextMenu ItemsSource="{Binding DbCommands}" DisplayMemberPath="Name">
-                            <ContextMenu.ItemContainerStyle>
-                                <Style TargetType="MenuItem">
-                                    <Setter Property="Command" Value="{Binding Source={StaticResource Proxy}, Path=Data.DbCommandSelected}" />
-                                    <Setter Property="CommandParameter" Value="{Binding}" />
-                                </Style>
-                            </ContextMenu.ItemContainerStyle>
-                        </ContextMenu>
-                    </ext:DropdownButton.DropdownMenu>
-                </ext:DropdownButton>
-
-                <Separator Margin="5,0,5,0" Style="{StaticResource {x:Static ToolBar.SeparatorStyleKey}}" />
-                
-                <Button Margin="0,0,5,0" Command="{Binding ParameterSettingsCommand}" CommandParameter="{Binding ElementName=queryEditor, Path=SelectedQuery}">
-                    <StackPanel Orientation="Horizontal">
-                        <Rectangle Height="18" Width="16" Fill="Black" >
-                            <Rectangle.OpacityMask>
-                                <VisualBrush Stretch="Fill" Visual="{StaticResource appbar_settings}" />
-                            </Rectangle.OpacityMask>
-                        </Rectangle>
-                        <TextBlock Margin="5,0,0,0" Text="Parameters" VerticalAlignment="Center" />
-                    </StackPanel>
-                </Button>
-            </ToolBar>
-        </ToolBarTray>
-
-        <teviews:TabbedQueryEditor x:Name="queryEditor" Grid.Row="1" />
-
-        <teviews:QueryExecutionResultsPanel Grid.Row="2" Results="{Binding Path=QueryExecutionStatisticsTable}" />
-
-        <Grid Grid.Row="3">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="Auto"/>
-                <ColumnDefinition Width="Auto"/>
-                <ColumnDefinition Width="*"/>
-            </Grid.ColumnDefinitions>
-
-            <Label Grid.Column="0" VerticalContentAlignment="Center" HorizontalContentAlignment="Left" Content="Mean Execution Time:"/>
-            <Label Grid.Column="1" VerticalContentAlignment="Center" HorizontalContentAlignment="Left" Content="{Binding AverageExecutionTime, StringFormat=hh:mm:ss.fff}"/>
-
-            <Label Grid.Column="2" VerticalContentAlignment="Center" HorizontalContentAlignment="Right" Content="{Binding Elapsed, StringFormat=hh:mm:ss.fff}" />
-        </Grid>
-    </Grid>
-</Page>
