+++ SqlQueryStressGUI/TestEnvironment/Views/TabbedQueryEditor.xaml.cs
@@ -1,5 +1,4 @@
 ﻿using SqlQueryStressGUI.Controls;
-using SqlQueryStressGUI.QueryStressTests;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
